feat(HomeScreen): changing the type of height and weight, from double to int and adding the unit to the title of the fields
- The height and weight type were changed to int, previously they were double.
- Added the unit in the field titles, height cm and weight kg.

+++ lib/src/models/bmi_data_model.dart
@@ -2,8 +2,8 @@ import 'package:flutter_bmi_calculator_dio/src/models/body_metrics_model.dart';
 
 class BmiDataModel extends BodyMetricsModel {
   BmiDataModel({
-    required double height,
-    required double weight,
+    required int height,
+    required int weight,
     required this.bmi,
     required this.interpretation,
   }) : super(
